Merged revisions 58638 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r58638 | russell | 2007-03-09 17:59:10 -0600 (Fri, 09 Mar 2007) | 8 lines

Merge some updates to the SLA documentation.  I plan to keep working on this
to explain all of the expected behavior with call handling, configuration
details for specific phones, and other things.  However, I got tired of doing
it in plain text, so I switched to using LaTeX.  I have included the PDF version.
I haven't been able to get a nice looking plain text version out of it yet, but
I'm not terribly concerned since this is supposed to be more of the manual,
while the plain text sample configuration file is the reference.

........