Blocked revisions 114813 via svnmerge
........
r114813 | mmichelson | 2008-04-28 17:38:07 -0500 (Mon, 28 Apr 2008) | 10 lines

Adding a new option 'n' to app_chanspy. This option allows for the name of the spied-on
party to be spoken instead of the channel name or number.

This was accomplished by adding a new function pointer to point to a function in app_voicemail
which retrieves the name file and plays it. This makes for an easy way that applications may play
a user's name should it be necessary. app_directory, in particular, can be simplified greatly by
this change.

This change comes as a suggestion from Switchvox, which already has this feature. AST-23

........