This commit compensates for buggy poll(2) implementations. Asterisk has, for a long time, had its own implementation of poll(2) which just used the input arguments to call select(2). In 1.4, this internal implementation was used for Darwin systems. This was removed in Asterisk trunk at some point, but it seems as though this was not the right move to make.
On Mac OS X, it appears as though the poll used
to gather CLI input does not respond properly
when connecting via a remote Asterisk console.
Reverting to the use of Asterisk's poll fixed
the issue.

Also, there is now an option for the configure
script, --enable-internal-poll, which will allow
for anyone to use Asterisk's internal poll
implementation in case they suspect that their
system's poll implementation is buggy.

closes issue #11928)
Reported by: adriavidal
Patches:
      1.6.0-configurev2.patch uploaded by putnopvut (license 60)