Fix underlinking problem in acinclude.m4 for newer gcc's.
Newer setups (distros) and gccs enforce -Wl,--as-needed and --no-undefined.
This means that the order of options to gcc are now important. In particular,
libraries must come after the object files that need them.

It is now wrong to pass -lfoobar into LDFLAGS. It should be passed *after*
the object, in LIBADD.

At this commit, the AST_EXT_TOOL_CHECK was fixed to have the libs after
the test program.
This fixes detection of the net-snmp library, at least.