Merged revisions 164272 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r164272 | russell | 2008-12-15 10:17:55 -0600 (Mon, 15 Dec 2008) | 8 lines

When a reload is issued, always process the configuration for dundi.conf.

The reason is that a reload can be used to refresh DNS lookups for defined peers.
Even if the config file hasn't changed, we want to process it for that purpose.

(closes issue #13776)
Reported by: kombjuder

........