Fix bridging regression from commit 176701
This fixes a bad regression where the bridge would exit after an attended
transfer was made. The problem was due to nexteventts getting set after the
masquerade which caused the bridge to return AST_BRIDGE_COMPLETE.

The problem is not completely solved however. A bridge configured with a warning
still does not operate correctly. It actually operates the same as before except
the DTMF is properly recognized after the first warning. I just wanted to get
this fix in as quickly as possible.

(issue #14315)
Reported by: tim_ringenbach