Allow bootstrap script to work on Solaris.
As usual, the way they do things is different, so we need to account for that.
automake is versioned ala BSD/Linux, but autoconf is not.  We don't actually
need to specify a version there, since AC_PREREQ will cover it for us.  Things
will fail pretty loudly if AC_PREREQ isn't met.

(closes issue #16341)
Reported by: bklang
Patches: 
      opensolaris_bootstrap.sh uploaded by bklang (license 919)