DTMF wasn't being logged on connected consoles when enabled in logger.conf
Previously in order for DTMF to be logged in a connected console session, the user would
have to do logger set channel DTMF on.  This corrects that so that it is on by default.
This issue was caused by an off by one error incurred by a logger level count of 6 in
logger.h where it should have been 7.

(closes issue: ASTERISK-17974)
Reported by: Luke H