res_config_pgsql: interpret "(null)" as NULL
in chan_sip we call this ugly ast_sockaddr_stringify_addr() which will
yield "(null)" for empty addresses, ports. This breaks the type casting
for integers, stores that garbage.

This is a temporary workaround in the lower layer. It should be fixed
in chan_sip instead, when all realtime engines can support NULL params.