Fixes issue with ConfBridge participants hanging up during DTMF feature menu usage getting stuck in conference forever.
When a conference user enters the DTMF menu they are suspended from the
bridge while the channel is handed off to the DTMF feature code.  If a
user entered this state and hungup, there existed a race condition where
the channel could not exit the conference because it was waiting on a
signal that would never arrive.  This patch fixes that, because it would
stupid for me to talk about the problem and commit a patch for something else.

(closes issue ASTERISK-18829)
Reported by: zvision