res/ari/resource_channels: Don't return allocation failure on failed function
If a function fails to execute, it is most likely due to one of two reasons:
(1) The function doesn't exist or can't be read from
(2) The function is dangerous and is restricted based on the user's permissions

Currently we return allocation failure, which is incorrect. This updates the
reason code to more accurately reflect why the request failed.

ASTERISK-24215