rest-api/api-docs/events.json: Remove non-compliant 'extends' attribute
Prior to the release of Swagger 1.2, the attribute 'extends' was being
promoted as a possible way to show that a particular object extends an existing
object. Instead, the Swagger specification went with the 'subTypes' attribute
in the base object. This patch removes the unsupported attribute; the object
that the offending objects proposed to extend already lists them in its
'subTypes' attribute.

ASTERISK-24300 #close
Reported by: Bradley Watkins