Blocked revisions 431620
While it may not be obvious, r431620 should not occur in Asterisk 13.
* We no longer set the SIP_DEFER_BYE_ON_TRANSFER flag during the handling of
  the INVITE with Replaces. This is now set and handled explicitly in the
  attended transfer and blind transfer code.
* An INVITE with Replaces replacing a channel in a Bridge will now safely eject
  the channel being replaced. No masquerade occurs.
* An INVITE with Replaces replacing a channel not in a Bridge will masquerade,
  but will do so in such a fashion that we can ensure that we are hanging up
  the channel when completed.

Since the code the patch fixes no longer exists due to core framework changes,
we should send a BYE naturally without the need for the flag.

........
channels/chan_sip: Ensure that a BYE is sent during INVITE w/Replaces transfer

Consider a scenario where Alice and Bob have an established dialog with each
other external to Asterisk. Bob decides to perform an attended transfer of
Alice to Asterisk. In this case, Alice will send an INVITE with Replaces
to Asterisk, where the Replaces specifies Bob's dialog with Asterisk. In this
particular scenario, Asterisk will complete the transfer, but - since Bob's
channel has had Alice masqueraded into it and is now a Zombie - a BYE
request will not be sent.

This patch fixes that issue by adding a new flag to chan_sip that tracks
whether or not we have an INVITE with Replaces. If we do, the flag is used
on the sip_pvt to ensure that a BYE request is sent, even if the channel has
been masqueraded away.

Review: https://reviewboard.asterisk.org/r/4362/

ASTERISK-22436 #close
Reported by: Eelco Brolman
Tested by: Jeremiah Gowdy, Kristian Høgh
patches:
  asterisk-11-hangup-replaced-3.diff uploaded by Jeremiah Gowdy (License 6358)